refactor(EatInOrder): 매장 주문 루트 어그리거트 기준으로 패키징하라
