+++ src/main/java/kitchenpos/eatinorders/order/application/OrderService.java
@@ -1,7 +1,11 @@
-package kitchenpos.eatinorders.application;
+package kitchenpos.eatinorders.order.application;
 
 import kitchenpos.deliveryorders.infra.KitchenridersClient;
-import kitchenpos.eatinorders.domain.*;
+import kitchenpos.eatinorders.order.domain.Order;
+import kitchenpos.eatinorders.order.domain.OrderLineItem;
+import kitchenpos.eatinorders.order.domain.OrderRepository;
+import kitchenpos.eatinorders.order.domain.OrderStatus;
+import kitchenpos.eatinorders.order.domain.OrderType;
 import kitchenpos.eatinorders.ordertable.domain.OrderTable;
 import kitchenpos.eatinorders.ordertable.domain.OrderTableRepository;
 import kitchenpos.menus.menu.domain.Menu;
