+++ src/main/java/kitchenpos/eatinorders/order/domain/JpaOrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.order.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
