+++ src/main/java/kitchenpos/eatinorders/ordertable/application/OrderTableService.java
@@ -1,7 +1,7 @@
 package kitchenpos.eatinorders.ordertable.application;
 
-import kitchenpos.eatinorders.domain.OrderRepository;
-import kitchenpos.eatinorders.domain.OrderStatus;
+import kitchenpos.eatinorders.order.domain.OrderRepository;
+import kitchenpos.eatinorders.order.domain.OrderStatus;
 import kitchenpos.eatinorders.ordertable.domain.OrderTable;
 import kitchenpos.eatinorders.ordertable.domain.OrderTableRepository;
 import org.springframework.stereotype.Service;
